refactor(workspace): decompose use-workspace into focused sub-hooks and modules
Replace 2570-line god hook with focused modules:
- workspace-types.ts: shared types and helpers
- workspace-session-store.ts: normalized session state
- use-workspace-sessions.ts: session CRUD/pagination/selection
- use-workspace-messages.ts: messages state/loading/refresh
- use-workspace-streaming.ts: SSE streaming, send/resume/abort
- use-workspace-model-selection.ts: models/agents/session model selection
- use-workspace-composed.ts: orchestrated composition
- __tests__/workspace-session-store.test.ts: store unit tests

Public API preserved via 10-line facade in use-workspace.ts.
Tests pass: 4100 passed (web), Podman images pass.
Lint passes: 0 errors, 29 warnings (13 new exhaustive-deps).
